load the defaults, cli, and local options in order
