+++ .gitignore
@@ -164,3 +164,4 @@ dummy.py
 gofile.env
 *.csv
 gofile-alljs-*
+md5_speed.py
