+++ README.md
@@ -12,6 +12,7 @@ Built using `asyncio`, `aiohttp`, and `tqdm`
 - Progress bars
 - Upload logging
 - Skipping duplicate uploads
+- Local configs
 
 ## Usage
 1. `pip install GofileIOUploader`
@@ -29,7 +30,7 @@ Gofile.io Uploader supporting parallel uploads
 positional arguments:
   file                  File or directory to look for files in to upload
 
-options:
+optional arguments:
   -h, --help            show this help message and exit
   -t TOKEN, --token TOKEN
                         API token for your account so that you can upload to a
@@ -42,19 +43,21 @@ options:
                         name if used
   -d, --dry-run         Don't create folders or upload files
   --debug-save-js-locally, --no-debug-save-js-locally
-                        Debug option to save the retrieved js file locally
+                        Debug option to save the retrieved js file locally.
+                        (default: False)
   -c CONNECTIONS, --connections CONNECTIONS
-                        Maximum parallel uploads to do at once
+                        Maximum parallel uploads to do at once. (default: 6)
   --public, --no-public
                         Make all files uploaded public. By default they are
-                        private and not unsharable
+                        private and not unsharable. (default: False)
   --save, --no-save     Don't save uploaded file urls to a
-                        "gofile_upload_<unixtime>.csv" file
+                        "gofile_upload_<unixtime>.csv" file. (default: True)
   --use-config, --no-use-config
                         Whether to create and use a config file in
-                        $HOME/.config/gofile_upload/config.json
+                        $HOME/.config/gofile_upload/config.json. (default:
+                        True)
   -r RETRIES, --retries RETRIES
-                        How many times to retry a failed upload
+                        How many times to retry a failed upload. (default: 3)
 
 ```
 ## Details
@@ -67,6 +70,44 @@ Configs are stored in `$HOME/.config/gofile_upload/config.json` and all successf
 Each time you complete an upload a `gofile_upload_<timestamp>.csv` will be created with the items uploaded and the following metadata:
 `filePath,filePathMD5,fileNameMD5,uploadSuccess,code,downloadPage,fileId,fileName,guestToken,md5,parentFolder`
 
+### Local Configuration
+A local configuration can be used when you specify the `--use-config` flag
+This config is saved in `$HOME/.config/gofile-upload/config.json` and contains the following options:
+```json
+{
+  "token": "Optional[str]",
+  "zone": "Optional[str]",
+  "connections": "Optional[int]",
+  "public": "Optional[bool]",
+  "save": "Optional[bool]",
+  "retries": "Optional[int]",
+  "history": {
+    "md5_sums": {
+      "<filepath>": "<md5sum of file>"
+    },
+    "uploads": [
+      "<upload response>"
+    ]
+  }
+}
+```
+This config is loaded at runtime and combined with CLI options and defaults to make one config for the program.
+The precedence for this is:
+1. CLI Defaults
+   - connections: 6
+   - public: False
+   - retries: 3
+   - save: True
+   - debug_save_js_locally: False
+   - use_config: True
+2. Local Config
+3. CLI Options
+
+The config will be saved when MD5 sums are calculated for files as well as when uploads are completed.
+Configs that have a value of `None` will be omitted.
+
+If you specify `--no-use-config` the local file will not be loaded and will not be saved to.
+
 ## Examples
 Given
 ```
@@ -74,25 +115,26 @@ directory/
 ├── sample2.mkv
 └── sample.mkv
 ```
-**Upload single file anonymously**
+**Upload single file anonymously** 
+The file will be private
 
 `gofile-upload directory/sample.mkv`
 
-**Upload single file to your account**
+**Upload single file to your account and make it public**
 
-`gofile-upload --token 123 foo directory/sample.mkv`
+`gofile-upload --token 123 --public directory/sample.mkv`
 
-**Upload single file to directory `foo` in your account**
+**Upload single file to directory `foo` in your account and make it public**
 
-`gofile-upload --token 123 --folder foo directory/sample.mkv`
+`gofile-upload --token 123 --public --folder foo directory/sample.mkv`
 
-**Upload directory to your account**
+**Upload directory to your account and make them public**
 
-`gofile-upload --token 123 directory`
+`gofile-upload --token 123 --public directory`
 
-**Upload directory to directory `foo` in your account**
+**Upload directory to directory `foo` in your account and make them public**
 
-`gofile-upload --token 123 --folder foo directory`
+`gofile-upload --token 123 --public --folder foo directory`
 
 # Development
 ## Optional Prerequesites
